readtty(): quick shot: work in pipelines (on non-TTY)..
The TTY layer will need *massive* work in the future, but to
allow a non-interactive thing like

   cat <<__EOT__ | s-nail -A work -~ -s cream
   Hey buddy, i'm attaching the work of today!
   ~@
   todays-work-utf8.txt
   utf-8

   .
   __EOT__

already in v14.0, we should catch invocations of readtty() with
is_a_tty[0] being false (seems to be a good thing anyway).
To make that smooth, simply read a line from STDIN in that case.