names.c:yankname(): fix quote regression..
(Rewrite *extract().., 2012-10-20) introduced a regression
regarding parsing of escaped quotes.  Given the example
  "x \"y\" z"
that would have resulted in
  "x"y" z"
Fix that.

But as documented in the message of the mentioned commit,
this is still intermediate in that we need real parsers
that conform to the RFC, and at the very best *only one*.
Until we have those this fix will hopefully suffice.