B[ase]64_ISTEXT fixes to overcome design pitfalls..
The current codebase is not prepared to deal with input lines
that form incomplete output lines, as can happen for at least
base64 encoded data due to the odd 3:4 relationship of input and
output (4:3 when decoding).

At the minimum we would have to deal with iconv(3) EILSEQ errors
and copy unconsumed data into temporary storage.
(Unfortunately our fault-tolerant iconv(3) wrapper skips over
EILSEQ just as it does for EINVAL, but this is a side note
only.)

The real solution requires detaching of decoding the
Content-Transfer-Encoding, the iconv(3) of the raw data, and the
consumation of the charset-converted data for display or other
uses of converted data (XXX mutt(1) offers a save-decoded).
Alternatively we must be prepared to handle leftover data on the
pre-CTE-decoded *and* the post-iconv() side.

This changeset adds a lot of TODO notes at several places in the
code, things that should be addressed for the send/MIME layer
rewrite that hopefully cumulates in S-nail v15.0.

It also changes handling of CONV_FROMB64_T cases in that we use
a heap buffer to store all the CTE-decoded data, which is the
only possible way to feed a correct multi-byte/-octet boundary
into (the) iconv() (of mime.c:fwrite_td()).
Like this even S-nail v14.0 will be able to display UTF-16 or
other multioctet base64-encoded data correctly;
it must be pointed out that this is still a huge improvement
compared to the Heirloom mailx(1) codebase, please see the
corresponding NEWS entry (enough thereof is contained in this
changeset).