mime.c:_fwrite_td(): rewrite..
This commit addresses two problems of _fwrite_td().

First, by adding the new _TD_BUFCOPY flag to enum tdflags, it
became possible to not copy over the buffer argument into
temporary storage unless there is a need to do so, and that
includes the case when the buffer is converted to temporary
storage anyway during execution of the function.
(Seems useful because in many cases the buffer argument is
itself a temporary dynamic buffer, one that may be adjusted on
the fly as necessary.)

Second, do not use alloca(3) based memory for the first of the
mentioned temporary buffers, but use heap even for that.  Given
that the current base64 decoder will pass the entire mimepart in
a single buffer -- and it'll do so until the MIME and send
layers have been rewritten -- it does not seem to be useful to
use stack-based storage for such, possibly huge, buffers.
Even though nail(1) (or at least Heirloom mailx(1)) always did
it like that ...