qp_encode(): do not waste one column..
When encoding body lines and a character that needs to be quoted
has to be written we've yet wasted a single column (column 76),
effectively forcing a soft line break after column 75.
So compare GT maximum not GE maximum instead.

Also, when in the last possible column after writing a character
that does not need to be quoted, do not enforce a soft line
break if we've written the LF or only an LF follows, as that'll
effectively end the line anyway.

Adjust the test checksums accordingly.