usermap(): add force_metoo argument..
The codebase is based upon a MUA that needs a MTA to do all the
user and address expansion etc., i.e., it assumes user map from
the systems user registry.  This is not true for S-nail, where
*from* etc. can be used to actually set the user identity as
such.

This means that an alias entry that matches the current user
name is fully expanded and not matched by any *metoo* clause,
even if that, in effect, doesn't mean a thing for the notion of
users of S-nail.

Adjust that in that *metoo* will be honoured for an alias entry
that matches the user name and that expands to a single entry
only.