Perform alias expansion for Reply-To: fields..
Martin Neitzel (neitzel AT escape DOT de) pointed out that alias
expansion is not performed for Reply-To: content, even though
that is a field that may contain a list of addresses on the one
hand and is definitely subject to changes due to *account*
switches on the other.

Unfortunately the codebase is a mess, so this is a rather
hackish implementation -- the changeset includes a lot of TODO
comments for how to perform better header setup, as to be
implemented in the MIME/send layer rewrite.

Btw., this is not covered by the POSIX standard, but as
Martin Neitzel pointed out correctly S-nail doesn't adhere the
standard *already*, since all of To:, Cc: and Bcc: are subject
to alias expansion, whereas POSIX only defines this for direct
recipients.
I think S-nail should be user friendly and continue to support
that, though.

This changeset disallows using of files etc. as Reply-To:
members etc.  This is *very* intermediate yet, though.