Bugfix for ~@ in non-interactive mode..
(readtty(): quick shot: work in pipelines (on non-TTY)..,
2013-01-25) introduced the possibility to "read data from the
terminal" (STDIN that is) in non-interactive mode.
Unfortunately i've never tested this like follows:

   $ cat <<_EOT | /s-nail -~ -s boom ./OUT
	~@
	test.c
	charset
	_EOT

Note the missing empty line after the two lines that specify the
filename and the charset of the attachment.

Well it turns out that i should have tested the readline()
return value for >0 instead of !=0, effectively ignoring the -1
End-Of-File condition, and tha in turn results in an endless
loop (could be that it even crashes if the memory allocator
randomizes addresses?  I think so.)