Hook *folder* update, add display name..
This can be seen as an improvement of (schdir(): realpath()
local files before leaving CWD.., 2013-01-08).

Add var_folder_updated() as a hook for the update of the
*folder* variable -- realpath(3) it right away as necessary and
if possible.  getfold() thus only returns a cached value.

Add the global *displayname* variable, and
lex.c:_update_mailname() to manage it as a size restricted
version of *mailname*, so that the user now sees "/BEGIN..END"
if that doesn't fit onto the display.

   Thanks to Christos Zoulas from the NetBSD project for the
   idea of separating the path and the user display!

With this new approach, schdir() can be slacked again.
And while here, fix some catgets() (and convert them to tr())
etc..