b64_decode_prepare(): fix CRAM-MD5 ,)..
Use spacechar() instead of whitechar() to trim the input,
and, voilá, IMAP steps on a bit.
It actually happens that the IMAP code passes the entire
response to the decoder, including the CRLF line terminator.
whitechar() does not match the CR, so that the decoder
failed right away since the input ends up non-aligned.
(IMAP login still broken.)