cram_md5_string(): ooops - fix it..
(aux.c: use new b64_*() family, 2012-11-17) turned to the new
Base64 CTE layer, but actually didn't take into account that the
original Heirloom code used "strtob64(sp)", i.e., worked on
"strlen(sp)", but the new code did *not* use b64_encode_cp(),
which would have been correct, but instead b64_encode(), and
that doesn't take care of strlen() but blindly trusts the length
of the string.  And *that* included the terminating NUL, so that
this was also encoded, and *that* effectively prevented logins.