Implement *datefield-markout-older* option..
Thanks to Stephen Isard we are now able to markout non-current
messages in a way that is comparable to what POSIX describes for
the -l option of the ls(1) command, i.e., messages that are
older than six months will use a different date format that is
easier to look at when searching in archives etc.

To avoid too many time(3) (sys)calls printhead() now requires
that time_current_update() is called before it's used (in the
current display cycle).

Note that also this change is somewhat intermediate, since the
_parse_head() execution path can surely be dropped at some
future time, in which case this option will work for *all*
possible dates, i.e., even for the From_ derived ones.
We yet do not have a parser that turns those addresses to
time_t values, and there are six or so different formats.
That's hard.  Maybe depend it upon strptime(3)?