smtp_auth_var(): fix user@addr specific lookups..
The temporary buffer was too small, so that the specific lookups
always failed due to buffer cuts (if the system offers
snprintf(); otherwise i think there will even be buffer
overflows).

I googled and see - this has been reported on the mailing
list, even with a patch!
So thanks go to jgw@txo.org, who reported this issue in november
2011.
And to Gavin Troy (gavtroy AT gmail DOT com), who sent in
a patch in january 2012.