Fix QP special soft linebreak message truncation..
(MIME: use b64_*() family for CONV_FROMB64* plus.., 2012-11-17)
was one of the  first steps to introduce sane error handling
etc. in send.c:sendpart().  (And still much missing btw.)

It happens, now that all the Content-Transfer-Encoding code has
been replaced, that a loop condition introduced in the mentioned
commit causes faulty truncation of quoted-printable encoded
message parts, if maliciousness (Apple Mail?  Hi Deborah
Goldsmith from Apple!) is happening, as in:

	One line.
	=
	This is in fact the second line.

The second line consists solely of a soft linebreak, resulting
in 0 written and 0 leftover bytes, a condition that i declared
erroneous back in november 2012.

Since it is possible (Apple Mail?) that a HTML part solely
consists of a single line, add an TODO note on that, since we
yet concatenate a (real) line until it is complete, which is
possibly not the real solution for these kind of mails ...