mime_fromhdr(): never return NULL output..
S-nail 14.1 can be crashed with an empty header field; i thought
this ends up as NULL/non-existent, but no, it doesn't.
Therefore we try to terminate a non-existent buffer, which ends
up as a SEGV.

Also, mime_fromhdr() *must* return with a *terminated*, *newly
allocated* buffer to satisfy its callers, so just do so.
This is XXX, of course.

Excuse:
	With the head at the Unicode stuff that i'm about to do
	i wanted to get that thing into the release, and obviously did
	a premature commit.  And here it is what happens.

While here, tweak code flow ++.