Add @ and @& shell command prefixes for pipe-MIMETYPE..
The former suppresses filters if multiple messages are displayed
at once, the latter adds asynchronous program execution on top of
that.
These should prevent denial-of-interaction DOI ,) if many
attachments require filters etc.
This may need additional tweaks, but we're early in the
development cycle.
While here, tweak some display messages.

I actually have to thank Gavin Troy (gavtroy AT gmail DOT com) who
stated that attachments were not properly working (which resulted
in S-nail release 14.2.2).
I did not know that people were really using the pipe- mechanism
as it was before (and even still is) because of the problems that
i tried to address with this commit.
Once we have a single pass parser we'll be able to ask the user
which MIME parts he actually wants to look at, but until then this
is a plaster.