print_collf() (~p): different layout, I...
First step to provide more informations about attachments in the
output of ~p.

For now we will only be able to provide Content-Type: information
if the file has a known MIME type, and charset information only if
either the user entered some or the MIME type indicated a non-text
format.  Furthermore -- today it may even happen that the actual
send will overwrite what we've shown or even fail (!) if automatic
character set conversion is impossible.

The problem is that our MIME file-content classifier will see the
file at a later time, which is already postponed until S-nail
15.0.  With the new MIME / send layer we'll classify the file at
once and thus be able to truly output the *final* Content-Type:
etc., and that will imply that the sending will not fail (due to
impossible conversion at least).

Nonetheless i think it's better to have some more info than to not
have it.