Drop *idna-strict-checks*..
Why should a MUA offer TLD checks like that when sending mail?
Also, this is very badly defined, e.g., LibIDN ships only data for
Norway (IIRC).
So save precious code space and remove the capability of
performing this check (the library ships the idn(1) utility, which
can be used to perform such a check as necessary).

As a nice side effect this fixes a possibly use-after-free bug
that i have done.