openssl.c:ssl_rand_init(): fix faulty expand/error check..
The commit a2b0daa5 from 2012-10-23 stated "This is a hairy
changeset, OpenSSL needs an eye.", but it seems the stupid author
wasn't yet in the right mood to perform proper review topic
branches, huuuuuhh???

Finally, almost seven (7) months later (!!!), fix that one and
print error messages if the file_expand() of *ssl-rand-egd* or
*ssl-rand-file* fails (other parts of the mentioned changeset did
that properly [the entire file is a mess] !?!?).

And also fix *smime-sign-include-certs-** again, which was
rendered unusable by this commit (bailing with error in case of
successfully expanding a file path).