Drop -I and -T options..
It seems Gunnar Ritter stopped developing nail/Heirloom mailx
once he started implementation of Newsreader functionality.
Well, for S-nail, it'll be a long time until we've iterated enough
over the code to reach the point where we can start to add such
big new features, and then we'll do that in a different way,
presumably.
So just get rid of what seems related to this Newsreader stuff.