Change smopts handling, mostly -r..
In main() we now have to deal with two dynamic char** arrays (one
for -S handling, one for the smopts), so unite their memory
management into a single function.

While there change the semantics of -r, which on the one hand
allows users much more control and freedom in what they are doing,
and on the other hand also allows us to simplify the code in
main().