imap.c: fix SIGSEGV on INTR..
	? fi imaps://user1@localhost
	Error with certificate at depth: 0
	 [.]
	 err 18: self signed certificate
	Continue (y/n)? y
	Password:Interrupt
	? set imap-auth=cram-md5
	? fi imaps://user1@localhost
	IMAP write error: error:140D00CF:SSL routines:SSL_write:protocol is shutdown
	Segmentation fault

Oh, my goodness.
The IMAP code is a terrific mess.

Here we yet called sclose() (after performing a siglongjmp(3) from
within a signal handler) for a local variable but didn't update
the global mb.mb_sock, which yet was assigned that variable, so
that on the next entry we assumed to be able to reuse a valid
connection.