`set folder=' tolerates `%:' and expands PROTOs PLUS..
  ? short xp %:imaps://user1@localhost
  ? set folder=xp

will set *folder* to `%:imaps://user1@localhost', as desired.
I.e., we now also explicitly ignore a `%:' prefix so that sharing
of strings is possible in resource files (as shown).

While here, modify vars.c:_check_special_vars() to return
a boolean so that assign() is capable to restore a value back to
the original version.

Use this in the special-var-hook var_folder_updated() to reject
tries to set *folder* to a POP3 protocol URI.