Drop expand_shell() (unused), change IMAP % expands..
in that `%' really means *folder* unless the IMAP URL therein
doesn't have a folder part, in which case `/INBOX' is appended;
before `%' always appended `/INBOX' for IMAP, though i'd already
documented the new behaviour.

While here, change expansion in var_folder_updated() from
EXP_SHELL to EXP_FULL.  This is more flexible.
And ensure that a we do not enter endless recursion if *MBOX* is
set to `&', which could happen due to this.
And treat an expanded `%' as final again, don't reenter expansion.

  In general all of this should be done in a different way; maybe
  once we have an URI object that parses anything for us, we could
  then decide what should be valid and what not.  I don't really
  know; i.e., i'm completely frustrated by the `@' special
  expansion, which, according to Gunnar Ritter, expands to the
  root of an IMAP *folder*, but **ONLY** if we are connected to
  that one, not otherwise!  Shouldn't this work relative to
  *folder* instead?  I'm afraid it should -- but, for this to
  work, the entire MUA needs to be rewritten, so that we can
  automatically connect SAFELY, which of course doesn't work right
  now, not even approximately...  This?  Maybe never.

Tweak documentation of "folder", *folder*, *MAIL* and *MBOX*.