At startup: fix mbstate & check wether enc has state..
The is-UTF-8 check at program startup may of course mess up the
encoding state, so FINALLY reset the state after this test.
This is the first step in TRULY supporting locking shift
encodings in S-nail.  (We'll turn to mbrtowc(3) all over.)

The good news is that the reset call also gives an indication
wether the character set encoding uses locking shift state
sequences or not.

Track this information via the new global enc_has_state, we need
it for the default builtin line editor.