execute(), cmdtab[]: fix leak, add `~' `call' alias..
[93ea8acd] added a ~MACRO `call' shortcut, but didn't add a real
command in cmdtab[], so do this now.

execute() yet used ac_alloc() to create a copy of the command name
(to be able to NUL terminate it), and even the mentioned commit
didn't change that.
This resulted in a memory leak if no alloca(3) is available,
because that buffer will be freed once a command returns regulary,
but not if an interruption or whatever else jump occurs during
command execution.  Fixed by using salloc() instead of ac_alloc().