prefixwrite(): temporarily fix for base64 4:3..
Well, i was pretty happy about the new approach as implemented in
(prefixwrite(): completely redefine *quote-fold*..), but a base64
encoded email from David Butenhof on austin-group-l@ revealed that
i had taken one thing NOT into accout: that prefixwrite() gets
called with whatever is there, which may be partial lines in the
base64 case: since S-nail 14.0 or so we are able to work it
sequentially: we've bitten ourselfs!

So this changeset overcomes this problem by introducing MORE
static cookies, so that we can take partial lines into account.
The result is as good as can be.  (Though David Butenhof's mail
revealed another problem, for which we yet have TODO entries,
and that is that the base64 layer doesn't strip \r from \r\n
sequences it sees, so as not to modify input data.  It yet doesn't
know wether that is allowed or not.)

The real solution has to wait until S-nail 15.0, when
prefixwrite() will become a caching object, so that it'll even be
able to pretty format entire paragraphs.  (If *quote-fold* is set
only, of course.)