makeconfig+: allow individual commline overwrites..
We've yet have had the restriction that `user.conf' had to be
adjusted if an option should have been changed, or the
CONFIG=CUSTOM approach had to be used but then `user.conf'
wouldn't have been read at all.

Throw away CONFIG=CUSTOM, always read `user.conf', but *only*
incorporate a `user.conf' value if the value doesn't yet exist.

This also allows to change the requirement that an option
explicitly has to be disabled to the opposite.