expand_shell_escape(): add \? escape sequence,..
document it under *prompt* and ensure execute() sets the new
global bool_t variable `exec_last_comm_error' to state wether the
last command failed or not.

Because `\?' is a S-nail specific extension, add a new
`use_nail_extensions' argument to expand_shell_escape(), so that
the other users are not bothered with it.
And change its behaviour in that unknown escape sequences are
echoed `as is' instead of being silently suppressed; like that our
`echo' produces identical output to bash(1), ksh(1) and dash(1),
only mksh(1) slightly differs.  (POSIX states

  The interpretation of a <backslash> followed by any other sequence
  of characters is unspecified.

.)  Therefore this changes (Add expand_shell_escape(),
2013-07-10), and also fixes it, because neither did we deal with
\\ correctly yet nor was a trailing backslash yet shown as such.