Enable CTRL-C during (non file-based) setfile()/quit()..
One of the reasons that we get stuck forever when the server
connection is lost during quit is a call to holdsigs() that is
placed in lex.c:setfile() right before calling quit().

This may be the right thing to do when acting on a MBOX mailbox,
but it's definitely wrong for all others, because those do not
really update anything local.
In fact they all do place a lot of signal and jump handlers just
as they should (well, if it's done like that, with blocking I/O).

So place a fat TODO note in the involved parties and hope that the
day will come when we have intermediate VOID boxes to jump to, so
that no crossing occurs (which would then isolate the holdsigs()
where needed, leaving an intermediate unprotected region).
PLUS.