Tweak krb5-config() and GNU make _buh..
First of all this tweaks (makeconfig: check GSSAPI via
krb5-config(1) if possible.., 2013-07-30) which didn't realize
that calling krb5-config(1) includes the currently set $CFLAGS
environment.  .-)

And then let's just make the `_buh' rule check wether we're setup
itself instead of calling the `all' rule because GNU make(1)
forgets that we've set CFLAGS= on the command line when it calls
recursively the `all' rule of the `mk.mk' makefile, therefore
loosing all the prescious compiler flags that we explicitly have
set in the outer call in that the CFLAGS= that occurs in `mk.mk'
takes precedence.  This is possibly even correct behaviour.