tty.c, NCL: fix SIGSEGV, optimize buffer takeover, +..
When ^R history searching a ^B or ^F history search entry,
a SIGSEGV occurred because the saved line content buffer
would be a NULL pointer.
Ensure to integrate neatly into the other history plus
movements while still being somewhat lean by temporarily
storing the savec buffer away, restoring it afterwards.
(Generally speaking we should use a real state machine,
but that would require more code, though it would also
allow us to support cursor movement, which would be a
good and desirable thing, i think.)

Optimize the buffer takeover case: instead of working char-wise,
simply span the entire buffer, effectively working character-wise.

Avoid some multiple-in-a-row beeps which occurred (and, dependent
on the code path, still occur) due to shared code.
Again, the way to overcome this would be to use a state machine,
which would allow much more, at the expense of more code.
(Note this change also simplifies the error path, but that was
a bitch anyway because of the lack of real error-handling support
in the ISO C multibyte handling.  On the other hand extend the
stack buffer to lower the probability of bailing due to redundant
shift states.  Ok, it is unlikely.)

And then document that ^G resets the multibyte state machine.
It's only half the truth, but at least we mention the word
"multibyte state machine" now somewhere.