tty.c, NCL: "avoid" excess of MAX_INPUT when tab-expanding..
The known problem of this code was yet that expansion may be so
large that the terminals idea of where the cursor is and what is
displayed is not identical to our own.
This is assumed to be a best-case scenario.  (May it be that old,
true consoles may even block?  Should try that out of interest.)

I left it like this consciously because the *real* solution would
be as is shown by true tab-*completion* enabled implementations:
show the *possible completions* page-wise as necessary.
This would however require a lot of logic, which is definitely
overkill at the current state of development.

However, given that the result looks flaky at best, and because we
cannot simply stop producing output when we would overflow the
possible maximum line length, let's just simply replace the
expansion with a message that says so.  It's one more TODO.