makefile: fix GNU make(1): drop .POSIX:..
today i've detected a misbehaviour of the new S-nail build system:
when using a command line like

  $ make WANT_QUOTE_FOLD=1 CFLAGS='-O2'

then GNU make(1) will not pass the CFLAGS= through to the
recursive make(1) process that performs the actual build.
Worse than that, it doesn't offer any possibility to get at the
command line arguments, so at the moment i fail to see
a workaround!

This is not a problem for the BSD make(1) and Schily make(1)
(smake(1)) -- these pass such arguments through to the build.
They both would also allow the workaround i had in mind, however,
because they do set the special `MAKEFLAGS' accordingly.  GNU
make(1) seems to simply interpret the POSIX wording

  The macro= value macro definition operands can also be included.

user unfriendly (note the `can').  (However, a shallow test on
UnixWare 7.1.4 made me believe that the UnixWare make(1) suffers
from the same problem.)

Right now i'm out of ideas how to proceed.

-----

Ha!  The above text was part of a message i was about to send to
S-nail(1) packagers and nail-devel@, when suddenly i had the idea
to simply drop the `.POSIX:' line from `makefile'.
And: it works!  (Oh, my goodness!)