Fix quotflt output channel (Gavin Troy)..
In [26b87285], (Anticipating v15.0: turn prefixwrite() into
object.., 2013-09-13)) i finally was able to get rid of that
terrible prefixwrite() with its static cookies.  This was
a preliminary change since of course the filter a.k.a. object
mechanism doesn't exist yet.

Therefore i fiddled this stream based object into the current send
mechanism.  The object gets re-initialized whenever sendpart() is
entered, which i thought was sufficient to get the output channel
targeted correctly.  I didn't realize, however, that, when working
multipart parts, the quotation filter output channel will not have
been reset, because there was no intervening sendpart() call and
thus no re-initialization ever occurred.

The super-safe fix is to always reset the quotflt to the original
output channel after sendpart() returned.
(The future goal is to plug a work-pipeline as necessary on
a per-part level, so that this shouldn't occur at first glance.
Unless there are bugs ;))