Collect mode: warn user if rfc822 attachments leave scope..
If the user uses `~:' (or `~_') to execute the `folder' command,
directly or indirectly, and there are rfc822 message attachments,
then those become either invalid or point to false messages.

Since this is practically impossible to fix with the current
codebase (with acceptable effort), at least ensure that we detect
the situation and warn the user if it happens.