cmd1.c:hprf(): *datefield-markout-older* (Stephen Isard)..
Stephen Isard checked against one week in the future once he
proposed the 'ls -l'-alike patch, but i've rejected that and went
truly with POSIX, which says that 'ls -l' must use the time-less
notation if the date is in the future.

It must be said that we should at least correct for the UTC-offset
of the mail we display, but as long as that is not implemented,
accept one day in the future from now on.