Fix 'test' target, and more (Gaetan Bisson)..
Gaetan Bisson correctly pointed out that 'make test' is actually
broken if the resulting binary is not ./s-nail, so fix that.

While here document that we do not support a $(SID)$(NAIL)
combination that contains whitespace (because that is not
make(1)-compatible), and remove any useless quotation marks that
gave that illusion.