tty.c, NCL: job control related fix..
S-nail was not reacting correctly when it was changed from
a background to a foreground process, or whatever causes the modes
of its terminal to change.
So, instead of querying the file-descriptor flags only once, query
them each and every time when attempting to change the mode.

(TODO If there were a global state machine for TTOU+, then it
TODO should be sufficient to requery it once we get one of these,
TODO plus after CONT, of course.)