Fix -u, change meaning of $USER to be EQ -u etc..
I broke -u a while ago.
That is bad, but handling of -u and $USER is terribly silly
all through the codebase anyway and in my opinion.
I.e., POSIX defines it so that it should only read the mailbox
of the given user, and NetBSD defines it that way.
However, the BSD codebase does in fact also use the "myname"
variable for elimination of alternates etc. etc.
This does not really make sense, so that we now do treat
-u so as that we *really* impersonate as the given user,
of course except that we do not switch the UID.

Behind the scenes the $USER environment variable has had
some influence, but rather silently.
This is also quite mysterious to me, and so i have decided
to change this to work just the very same as -u.

All of this means that we now have two systemcalls if $USER
is set (or -u is given), so as to decide wether we do incarnate
as someone else, but on the other hand we handle a bit more
correct than we did before.
*Because* we do not silently use some values here and there.

P.S.: while here, change main.c to no longer use magic constants
for ~/.mailrc and /tmp, and change main.c to s-it-mode.