tty.c:NCL:_ncl_kht(): TWEAK Linux tty interaction, plus..
On Linux the MAX_INPUT is truly 255 in that dropbear(1) and plain
TTY sessions get in real trouble if the linesize is exceeded, in
that nothing visual appears at all.

So take more care and clamp the error-replacement output more and
more until nothing but a string [ERR] remains.

Yet another real problem was that the code did yet not take any
care about the *prompt*, so that the "does-it-fit-onto-line" check
was complete nonsense!!  Since we simply put it via fputs(3) all
we can do is assume each byte in *prompt* requires two columns.

Finally, to be able to print at least some small error message if
all else fails, clamp the prompt to an allowed maximum.