Add srelax(_hold|_rele)?()..
sreset() drops all the auto-reclaimed storage and thus cannot
be used when iterating over all messages in a mailbox, because
there may have been allocations in that storage somewhere before.
Therefore introduce a temporary additional layer on top of that
which can be used by such code paths without resetting yet
existing (non-persistent, i.e., per-cycle) allocations.

This results in dramatical benefits, especially when changing the
flag of a message so that the mailbox needs to be updated, the
more messages there are in a mailbox, and the more use of MIME in
header fields is present, the more dramatical.