Build system: detect/set CC and CFLAGS, _buh -> devel, plus..
Detect our CC environment and try to set CC and CFLAGS to what we
think is best.  Packagers etc. could now add what they need via
passing ADDCFLAGS= and ADDLDFLAGS=.
However, to not rebuff users, only do so if CC, CFLAGS and LDFLAGS
is not set or set to the empty string, respectively.
Or -- if such action has been explicitly requested by the new
WANT_AUTOCC.

Also, let WANT_DEBUG force WANT_NOALLOCA.

One nice side effect of this is that the S-nail build system now
also recognizes changes in build flags etc., and forces
configuration updates and recompilation as necessary.

While here i checked usage with old Bourne shells and in effect it
didn't work because we've assumed that 'eval's from within a loop
with redirection would export their variable setting, but that
isn't true for Heirloom sh(1) and Schily sh(1).
Therefore we need to go through the redirection of a temporary
file, that we then reinclude via the . source operator.

Derive the paths of the standard utilities we need via 'command
-vp', and use those paths for the entire build system (no longer
anything in `conf.rc').  That turned out to fix UnixWare 7.1.4
installation, because the install(1) program there doesn't copy
a file onto a file (never tried, ooops).