acmava.c:_check_special_vars(): SIGBUS for *line-editor-cursor-right*..
we yet were not prepared for setting this variable to the empty
string (trying to terminate a .RODATA buffer).

The real problem however is that i almost right from the start
(end 2012 anyway) wanted to implement a value cache which would
define clear actions on values (and getting rid of string lookup);
that would or WILL also allow to forbid that a value is set to the
empty string.  But still, not existing.  8-/