Thread subjects: improve visibility situation..
Until now we simply didn't repeat any Subject: in thread followup
messages, which was a bad thing when a thread has spanned several
display pages.

Try to improve this situation ALREADY TODAY by counting the number
of messages that have already been displayed in the current output
run.  Seems to work at first glance, so MUCH better would be
TODO a nice output encapsulation and the possibility to simply have
TODO a routine which says either how many messages are on the display
TODO or gives the distance to the message which is on the top of the
TODO display.

This changeset introduces a new exported print_headers() function
and instead internalizes the sofar exported printhead() one.
Move the (now) _print_head() related static utility functions to
a continuous block etc. etc.