Allow overwriting of STRIP= (and all other tools) (Ypnose)..
Ypnose reported a cross-compilation issue on Void Linux [1],

  /bin/strip: Unable to recognise the format of the input file
  `/destdir/arm-linux-gnueabihf/s-nail-14.5.1/usr/bin/mail'

  [1] <https://github.com/voidlinux/xbps-packages/pull/238>

So introduce a check_tool() function instead of blindly using
command(1), and allow users to overwrite each and every tool we
need.

Also noted is that calling 'command -pv NAME' is an addition of
POSIX 2008 (Issue 7), though usable on all tested systems.