tty.c,NCL: FIX and truly implement in-memory limit (Gavin Troy)..
Gavin Troy reported a crash with [next] that i could track down to
the addition of the new `history' command: the reason was that the
current in-memory counter was not used at all before that, and it
seems i stopped thinking about at the time i've implemented the
NCL once i've realized that i could go with the maximum-cutlimit
only.  (In fact i think i should leave this code alone for five
months and then do a real review.)

Anyway, when adding the `history' command i started using the
in-memory limit without reviewing the rest of the code wether that
is really capable to deal with that, and that was the problem that
Gavin had to face on [next].  Ugh.

Worse: the fix i've first implemented, and which is fixup'd with
this changeset, was completely insufficient.

So do it for real, and implement a true in-memory limit for the
NCL, dropping oldest entries once the limit overflows.

Many, *many* thanks go to the brave Gavin Troy for facing [next]!