Bugfix [a9b67e9] (sort(),thread(): use srelax()!, 2014-01-18)..
Just as mentioned in the release notes (sic) that was a hasty
commit of an untested diff that i've added few minutes beforehand.

It was of course wrong, and i've placed a reverse patch on the
servers:

  http://downloads.sourceforge.net/project/s-nail/s-nail-14.5.2-sort.diff

The problem (no memory relaxation during entire sort operation) requires
a different approach (either only relax when we don't need to hold
dope storage (as for SORT_TO/SORT_FROM/SORT_SUBJECT/+) or use heap
memory for those allocations, then).

Using heap memory and keep using string dope relaxation is what
this changeset does, and i've also placed this slightly tested
forward-headed patch on the servers:

  http://downloads.sourceforge.net/project/s-nail/s-nail-14.5.2-sort-alt.diff