Add xz(1) compression, drop auto-detection of compressed boxes..
All of this TERRIBLE MESSY STUFF requires a lot of object-based
rework, but in the meanwhile drop automatic detection of
compressed box equivalences -- if a user wants to open a mbox.gz
she should just `fi mbox.gz'.

I took a deeper look when i lightly tested the added support of
xz(1) compression, as it revealed that all this time users of the
auto-detection had to face a `Can't canonicalize' warning message.
(`All this time': since i've added realpath(3) expansion to avoid
the endless-stuck-when-opened-folder-disappears problem that
unfortunately is a vivid part of the BSD Mail codebase.)

I don't feel to bad about this, though, since the *newmail*
mechanism NEVER (sorry for the noise) was capable to detect
changes on such auto-expanded boxes anyway.
And because of the ridiculous which_protocol() over-and-over-again
expensiveness it's much better to drop all that shit anyway.