lex.c:_update_mailname(): continue if realname() fails..
Instead of simply returning false perform the usual work of
copying the (failed-to-resolve) path to ::mailname and
::displayname.

The maildir code sofar chdir(2)d into the new maildir directory
before calling initbox(), so that the realpath was invoked from in
there and naturally failed due to that.  With this change we will
still print a "cannot canonicalize" warning, but at least will be
able to access the maildir folder correctly.
It had been broken for a long time -- ooops.