Change `[?..]?' search expression to `[@..]@'..
Like that it can be used on an empty line.
I really thought about doing so from the beginning, but
finally went for `?' because `@' is so uncommon for searches.
For S-nail, however, only the latter really fits, because of
course `?' clashes with the help command (if used on a line by
itself).